test: move self-anchor integration test
