+++ tests/networks/basic-rust.yaml
@@ -16,6 +16,7 @@ spec:
             CERAMIC_ONE_EVM_RPC_URL: "http://ganache:8545"
             CERAMIC_ONE_EVM_PRIVATE_KEY: "0000000000000000000000000000000000000000000000000000000000000001"
             CERAMIC_ONE_EVM_CHAIN_ID: "1337"
+            CERAMIC_ONE_ANCHOR_INTERVAL: "20"
             CERAMIC_ONE_EVM_CONTRACT_ADDRESS: "0x231055A0852D67C7107Ad0d0DFeab60278fE6AdC"
             CERAMIC_ONE_FLIGHT_SQL_BIND_ADDRESS: "0.0.0.0:5102"
           resourceLimits:
